requirements.txt
gspread==6.1.4
numpy==2.1.3
oauth2client==4.1.3
openpyxl==3.1.5
pandas==2.2.3
protobuf==5.28.3
psycopg2_binary==2.9.9
PyDrive==1.3.1
python-dotenv==1.0.1
tzlocal==5.2
